backend/requirements.txt
asgiref==3.7.2
autopep8==2.3.2
Django==5.0.11
djangorestframework==3.15.2
drf-yasg==1.21.7
psycopg2-binary==2.9.9
pycodestyle==2.12.1
sqlparse==0.4.4
typing_extensions==4.9.0
